move discardable op interfaces from transforms lib to dialect lib
