+++ include/CMakeLists.txt
@@ -38,6 +38,7 @@ add_subdirectory(llzk/Dialect/POD/IR)
 add_subdirectory(llzk/Dialect/POD/Transforms)
 add_subdirectory(llzk/Dialect/Polymorphic/IR)
 add_subdirectory(llzk/Dialect/Polymorphic/Transforms)
+add_subdirectory(llzk/Dialect/Shared)
 add_subdirectory(llzk/Dialect/SMT/IR)
 add_subdirectory(llzk/Dialect/String/IR)
 add_subdirectory(llzk/Dialect/Struct/IR)
