+++ include/llzk/Dialect/Shared/DiscardableAllocationOpInterfaces.td
