+++ lib/Dialect/Array/CMakeLists.txt
@@ -5,9 +5,20 @@ target_link_libraries(LLZKAllDialects INTERFACE LLZKArrayDialect)
 file(GLOB_RECURSE LLZKArrayDialect_SOURCES "**/*.cpp")
 target_sources(LLZKArrayDialect PRIVATE ${LLZKArrayDialect_SOURCES})
 target_link_libraries(
-  LLZKArrayDialect PUBLIC LLZKDialectHeaders ${LLZK_DEP_DIALECT_LIBS} MLIRIR
-                     MLIRPass MLIRParser MLIRTransformUtils MLIRSCFTransforms
-                     LLVMHeaders MLIRHeaders LLZKDialect)
+  LLZKArrayDialect
+  PUBLIC
+    LLZKDialectHeaders
+    LLZKDialectShared
+    ${LLZK_DEP_DIALECT_LIBS}
+    MLIRIR
+    MLIRPass
+    MLIRParser
+    MLIRTransformUtils
+    MLIRSCFTransforms
+    LLVMHeaders
+    MLIRHeaders
+    LLZKDialect
+)
 llzk_target_add_mlir_link_settings(LLZKArrayDialect)
 
 install(TARGETS LLZKArrayDialect EXPORT LLZKTargets)
