+++ lib/Transforms/CMakeLists.txt
@@ -6,8 +6,6 @@ file(GLOB LLZKTransforms_SOURCES "*.cpp")
 target_sources(
   LLZKTransforms PRIVATE
     ${LLZKTransforms_SOURCES}
-    DiscardableAllocationOpInterfaces.cpp
-    LLZKWhileToForPass.cpp
 )
 
 target_link_libraries(
